requirements.txt
streamlit
firebase-admin
llama-index
python-dotenv
names 